Revert "refactor common code"
This reverts commit 548804c9783a96d80bfd2bf20775de5a015c7436.
